src/modules/planning/use-cases/tasks/archive-task/archive-task.use-case.ts
import { Injectable, Logger } from '@nestjs/common';
import { AppErrors, Either, left, Result, right, UseCase } from '../../../../../shared/core';
import { Member } from '../../../domain/member.entity';
import { Task } from '../../../domain/task.entity';
import { TaskRepository } from '../../../domain/task.repository';

type Request = {
  member: Member;
  task: Task;
};

type Response = Either<AppErrors.UnexpectedError | Result<any>, Result<Task>>;

@Injectable()
export class ArchiveTaskUseCase implements UseCase<Request, Response> {
  private readonly logger = new Logger(ArchiveTaskUseCase.name);

  constructor(private readonly taskRepository: TaskRepository) {}

  async execute(request: Request): Promise<Response> {
    this.logger.log('Archiving task...');

    const { member, task } = request;

    try {
      if (task.isArchived()) {
        this.logger.debug('Task already archived.');
        return right(Result.ok(task));
      }

      const result = task.archive(member.ownerId);
      if (result.isFailure) {
        return left(result);
      } else {
        await this.taskRepository.save(task);
        this.logger.log('Task successfully archived.');
        return right(Result.ok<Task>(task));
      }
    } catch (error) {
      this.logger.error(error.message, error);
      return left(new AppErrors.UnexpectedError(error));
    }
  }
}
